refactor: Revert AssetUtil package move to avoid breaking change (#468)
