+++ core-api/src/main/java/com/bloxbean/cardano/client/api/model/Utxo.java
@@ -1,6 +1,6 @@
 package com.bloxbean.cardano.client.api.model;
 
-import com.bloxbean.cardano.client.transaction.util.AssetUtil;
+import com.bloxbean.cardano.client.api.util.AssetUtil;
 import com.bloxbean.cardano.client.transaction.spec.Asset;
 import com.bloxbean.cardano.client.transaction.spec.MultiAsset;
 import com.bloxbean.cardano.client.transaction.spec.Value;
