+++ core-api/src/test/java/com/bloxbean/cardano/client/api/util/AssetUtilTest.java
@@ -2,7 +2,6 @@ package com.bloxbean.cardano.client.api.util;
 
 import com.bloxbean.cardano.client.transaction.spec.Asset;
 import com.bloxbean.cardano.client.transaction.spec.MultiAsset;
-import com.bloxbean.cardano.client.transaction.util.AssetUtil;
 import com.bloxbean.cardano.client.util.HexUtil;
 import org.junit.jupiter.api.Test;
