+++ function/src/test/java/com/bloxbean/cardano/client/function/helper/SignerProvidersTest.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.client.function.helper;
 
 import com.bloxbean.cardano.client.account.Account;
-import com.bloxbean.cardano.client.transaction.util.AssetUtil;
+import com.bloxbean.cardano.client.api.util.AssetUtil;
 import com.bloxbean.cardano.client.common.model.Networks;
 import com.bloxbean.cardano.client.crypto.KeyGenUtil;
 import com.bloxbean.cardano.client.crypto.SecretKey;
