+++ quicktx/src/main/java/com/bloxbean/cardano/client/quicktx/blueprint/extender/MintValidatorExtender.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.client.quicktx.blueprint.extender;
 
 import com.bloxbean.cardano.client.api.model.Amount;
-import com.bloxbean.cardano.client.transaction.util.AssetUtil;
+import com.bloxbean.cardano.client.api.util.AssetUtil;
 import com.bloxbean.cardano.client.exception.CborRuntimeException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.plutus.blueprint.model.Data;
